+++ templates/next-image/package.json
@@ -11,10 +11,10 @@
   "dependencies": {
     "@ai-sdk/openai": "^2.0.16",
     "@merit-systems/echo-next-sdk": "0.0.14",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
-    "@radix-ui/react-select": "^2.0.0",
-    "@radix-ui/react-slot": "^1.0.2",
-    "ai": "^5.0.17",
+    "@radix-ui/react-dropdown-menu": "^2.1.16",
+    "@radix-ui/react-select": "^2.2.6",
+    "@radix-ui/react-slot": "^1.2.3",
+    "ai": "^5.0.40",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "^0.263.1",
