+++ templates/next-image/src/app/layout.tsx
@@ -14,7 +14,7 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
+  title: 'Echo Image Gen',
   description: 'Generated by create next app',
 };
