+++ templates/next-image/src/app/page.tsx
@@ -19,7 +19,7 @@ export default async function Home() {
       <header className="flex justify-between items-center w-full mb-8 p-6 bg-gradient-to-r from-slate-50 to-gray-100 rounded-xl border border-gray-200 shadow-sm">
         <div className="flex items-center space-x-3">
           <h1 className="text-3xl font-mono bg-gradient-to-r from-gray-900 to-gray-600 bg-clip-text text-transparent">
-            Next.js Image Generation Example
+            Echo Image Gen
           </h1>
         </div>
         <div className="flex items-center gap-3">
